# SWAGGER_API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: Lilium B2B E-commerce Backend API
  description: Complete API documentation for the Lilium B2B multi-vendor platform with cash-on-delivery system
  version: 1.0.0
  contact:
    name: Lilium API Support
    email: support@lilium.iq
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.lilium.iq/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
      responses:
        200:
          description: List of orders

    post:
      tags: [Orders]
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, deliveryAddress, zone]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                deliveryAddress:
                  type: string
                zone:
                  type: string
                notes:
                  type: string
                promotionCode:
                  type: string
      responses:
        201:
          description: Order created
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order details
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ACCEPTED, PREPARING, ON_THE_WAY, DELIVERED, CANCELLED]
      responses:
        200:
          description: Status updated
...
  /vendors/orders:
    get:
      tags: [Vendors]
      summary: List vendor orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        200:
          description: Vendor orders
...